separated out settings component
