+++ src/components/SettingsRadioButton.tsx
@@ -1,15 +1,19 @@
 const { React } = window;
 import { THEME } from "../styles";
 
+const style: React.CSSProperties = {
+  border: "2px solid black",
+  borderRadius: "5px",
+  margin: "5px"
+};
+
 export default function SettingsRadioButton(
   {current, target, label, onClick}:
   {current: number, target: number, label: string, onClick: (e: number) => void}
 ) {
   return <button
     style={{
-      border: "2px solid black",
-      borderRadius: "5px",
-      margin: "5px",
+      ...style,
       backgroundColor: current === target ? THEME.midLight : THEME.midDark
     }}
     onClick={() => onClick(target)}
