+++ src/styles.ts
@@ -4,7 +4,7 @@ export const THEME = {
   dark: "#0D1321",
   midDark: "#7F838B",
   midLight: "#B7BBC0",
-  light: "#F0F0F1",
+  light: "#FCFCFC",
 } as const;
 
 export const TEXT_SIZES = {
@@ -44,42 +44,6 @@ export const GLOBAL_STYLES = {
     fontSize: "25px",
     width: "100%"
   },
-  settings: {
-    window: {
-      backgroundColor: THEME.light,
-      border: "2px solid black",
-      display: "flex",
-      flexDirection: "column",
-      flex: 9,
-      overflow: "auto",
-      position: "relative",
-      width: "100%"
-    },
-    header: {
-      alignItems: "center",
-      display: "flex",
-      justifyContent: "center",
-      minHeight: "3em",
-      margin: "10px",
-      position: "relative"
-    },
-    row: {
-      alignItems: "center",
-      display: "flex",
-      flexDirection: "row",
-      height: "1.5em",
-      minHeight: "3em",
-      padding: "1em"
-    },
-    label: {
-      position: "absolute",
-      left: "5px"
-    },
-    parameter: {
-      position: "absolute",
-      right: "5px"
-    }
-  },
   tabContainer: {
     alignItems: "end",
     display: "flex",
@@ -90,21 +54,23 @@ export const GLOBAL_STYLES = {
   },
   tab: {
     border: "2px solid black",
-    borderBottom: "0px solid black",
-    borderTopLeftRadius: "8px",
-    borderTopRightRadius: "8px"
+    borderBottom: "none",
+    borderTopLeftRadius: "5px",
+    borderTopRightRadius: "5px"
   },
   smoothContainer: {
     alignItems: "center",
     backgroundColor: THEME.light,
-    borderBottom: "1px solid black",
+    borderBottom: "2px solid black",
     display: "flex",
-    height: "3em",
+    height: "5vh",
     justifyContent: "start",
     padding: ".5em"
   },
   numberInput: {
     backgroundColor: THEME.light,
+    border: "2px solid black",
+    borderRadius: "5px",
     height: "2ch",
     padding: "5px",
     textAlign: "right",
@@ -146,54 +112,53 @@ export const GLOBAL_STYLES = {
     overflowY: "scroll",
     width: "100%"
   },
-  trigger: {
-    container: {
-      alignItems: "start",
-      border: "1px solid black",
-      display: "flex",
-      flexDirection: "column",
-      padding: "12px",
-      position: "relative"
-    },
-    buttonContainer: {
-      alignItems: "center",
-      display: "flex",
-      flexDirection: "row",
-      direction: "ltr",
-      justifyContent: "space-between",
-      position: "absolute",
-      right: "5px"
-    },
-    property: {
-      alignItems: "center",
-      display: "flex",
-      flexDirection: "row",
-      justifyContent: "start",
-      marginBottom: "0.25em",
-      width: "100%",
-      whiteSpace: "nowrap"
-    },
-    text: {
-      margin: "0em .5em"
-    },
-    createButton: {
-      alignItems: "center",
-      backgroundColor: THEME.light,
-      bottom: "-0.75em",
-      border: "2px solid black",
-      display: "flex",
-      fontSize: "18px",
-      height: "1.5em",
-      justifyContent: "center",
-      left: "0px",
-      marginLeft: "auto",
-      marginRight: "auto",
-      position: "absolute",
-      right: "0px",
-      userSelect: "none",
-      width: "1.5em",
-      zIndex: 1
-    }
+  triggerContainer: {
+    alignItems: "start",
+    borderBottom: "2px solid black",
+    display: "flex",
+    flexDirection: "column",
+    padding: "12px",
+    position: "relative"
+  },
+  triggerActionContainer: {
+    alignItems: "center",
+    display: "flex",
+    flexDirection: "row",
+    direction: "ltr",
+    justifyContent: "space-between",
+    position: "absolute",
+    right: "5px"
+  },
+  triggerPropContainer: {
+    alignItems: "center",
+    display: "flex",
+    flexDirection: "row",
+    justifyContent: "start",
+    marginBottom: "0.25em",
+    width: "100%",
+    whiteSpace: "nowrap",
+  },
+  propLabel: {
+    margin: "0em .5em"
+  },
+  newTriggerButton: {
+    alignItems: "center",
+    backgroundColor: THEME.light,
+    bottom: "-0.75em",
+    border: "2px solid black",
+    borderRadius: "5px",
+    display: "flex",
+    fontSize: "18px",
+    height: "1.5em",
+    justifyContent: "center",
+    left: "0px",
+    marginLeft: "auto",
+    marginRight: "auto",
+    position: "absolute",
+    right: "0px",
+    userSelect: "none",
+    width: "1.5em",
+    zIndex: 1
   },
   dropdown: {
     head: {
